refactor: 상수 파일을 utils 패키지로 이동
